Extract out data table types
