+++ editor.planx.uk/src/ui/shared/DataTable/components/icons.tsx
@@ -1,0 +1,7 @@
+import Close from "@mui/icons-material/Close";
+import Done from "@mui/icons-material/Done";
+import React from "react";
+
+export const True: React.FC = () => <Done color="success" fontSize="small" />;
+
+export const False: React.FC = () => <Close color="error" fontSize="small" />;
